Work around the cold start problem by falling back to rules
We need good training data to create a high quality classifier from the
user's data. If the data set does not contain enough important messages,
the ML would learn that few things are important and thus always develop
a bias towards non-important classification. To solve this we fall back
to a lower quality rule-based classifier that uses the same features as
the ML classifier. The importance is determined by hard-coded thresholds
of the individual features.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>